[ALPS07845127] mml: rrot bin odd
Fix RROT binning odd rect by calculate round width and height,
pass to down stream engines before rsz. And rsz uses frame crop
with sub pixels.
Fix overflow issue in rsz fw.

MTK-Commit-Id: 8748ff77066a4b9edb9d8f4b9f5804c8730bad2a

CR-Id: ALPS07845127
Feature: mml

+++ drivers/gpu/drm/mediatek/mml/mtk-mml-rsz-fw.c
@@ -78,7 +78,7 @@ static void rsz_config(struct rsz_fw_in *in, struct rsz_fw_out *out,
 	u32 crop_subpix_w = in->crop.w_sub_px;
 	u32 crop_height = in->crop.r.height;
 	u32 crop_subpix_h = in->crop.h_sub_px;
-	s32 alg;
+	u32 alg;
 
 	if (is_hor)
 		alg = out->hori_algo;
@@ -216,13 +216,13 @@ static void rsz_auto_align(struct rsz_fw_in *in, struct rsz_fw_out *out,
 	bool is_hor, struct rsz_cal_param *cal_param)
 {
 	s32 prec, max_nm, max_nm_prec, offset = 0;
-	u32 dst_width = in->out_width;
-	u32 dst_height = in->out_height;
-	u32 crop_width = in->crop.r.width;
-	u32 crop_subpix_w = in->crop.w_sub_px;
-	u32 crop_height = in->crop.r.height;
-	u32 crop_subpix_h = in->crop.h_sub_px;
-	s32 alg;
+	s64 dst_width = in->out_width;
+	s64 dst_height = in->out_height;
+	s64 crop_width = in->crop.r.width;
+	s64 crop_subpix_w = in->crop.w_sub_px;
+	s64 crop_height = in->crop.r.height;
+	s64 crop_subpix_h = in->crop.h_sub_px;
+	u32 alg;
 
 	if (is_hor)
 		alg = out->hori_algo;
